requirements.txt
google-auth~=1.28.0
google-cloud-secret-manager~=2.3.0
pytest~=6.2.2